Update windows.md
In case you want to show the command line. The docs do not specify how to install for WSL2. And it was easy to use the wrong Linux command line and to dive into a need to set up systemd (which is not included so far with WSL), which does not seem to be required for IBM Cloud CLI. The CLI Docs also mention Windows Pro, which installs in PowerShell and not in WSL.